CON-478: Moved the bastard merge to tests.
